Specification: Replace forms with their /key variants.
Each procedure that previously had a /key variant is now equivalent
to that variant, that is, the value-only versions of these forms have
been removed.

While I believe that having both value and key+value versions of forms
like imapping-map is convenient, the presence of all these /key forms
was clearly an idiosyncracy of this SRFI and decreased its
compatibility with SRFI 146 and other libraries unnecessarily.