Rename imappings to fxmappings and restrict keys to fixnums.
The "i" prefix has already been used to denote immutable structures in
SRFIs 116, 134, and 140.  Thus, the name "imapping" was ambiguous.

This change also restricts keys of integer mappings to the fixnums;
performance and compatibility concerns make the optional extension
to bignums undesirable.

For the discussion leading up to this change, see the mailing-list
thread starting with
https://srfi-email.schemers.org/srfi-224/msg/16814545/