Remove fxmapping-partition-map.
While this function has been in the Haskell IntMap library for a long
time, it is half-baked in its current form.  As pointed out on the
mailing list, the restriction to passing two continuations seems
arbitrary, perhaps an artifact of the use of Either.  Figuring the
most general, useful form of this function can be a project for the
future.